Move original gribmagic implementation to gribmagic.unity
