+++ .gitignore
@@ -5,7 +5,6 @@ __pycache__
 
 .gribmagic-data
 .gribmagic-testdata
-tests/modules/download/fixtures
 
 tools/dwd-grib-downloader
 tmp
