+++ docs/publishing/xpublish.rst
@@ -35,8 +35,8 @@ Server
 ======
 ::
 
-    export BASE_STORE_DIR=.gribmagic-data
-    python demo/publishing/xpublish_server.py $BASE_STORE_DIR/tmp/icon_eu_20201220_00_air_temperature_2m_000.grib2
+    export GM_DATA_PATH=.gribmagic-data
+    python demo/publishing/xpublish_server.py $GM_DATA_PATH/tmp/icon_eu_20201220_00_air_temperature_2m_000.grib2
     open http://localhost:9000/
