+++ gribmagic/unity/enumerations/weather_models.py
