--- tests/conftest.py
@@ -1,6 +1,0 @@
-import os
-
-os.environ["BASE_STORE_DIR"] = "/app/data"
-os.environ["MODEL_CONFIG"] = "config/model_config.yml"
-os.environ["MODEL_VARIABLES_MAPPING"] = "config/model_variables_mapping.yml"
-os.environ["MODEL_VARIABLES_LEVELS_MAPPING"] = "config/model_variables_levels_mapping.yml"
