+++ tests/unity/conftest.py
@@ -1,0 +1,3 @@
+import os
+
+os.environ["GM_DATA_PATH"] = "/app/data"
